[WIP] Before merge of product_attribute_priority
