+++ product_code_builder/models/__init__.py
@@ -15,4 +15,8 @@
 #    along with this program.  If not, see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from . import product
+
+from . import product_attribute_value
+from . import product_product
+from . import product_template
+
