+++ product_code_builder/views/product_attribute_value_view.xml
