+++ product_code_builder/views/product_view.xml
